+++ inmemory/src/main/java/org/vertexium/inmemory/InMemoryVertex.java
@@ -8,9 +8,9 @@ import org.vertexium.query.VertexQuery;
 import org.vertexium.search.IndexHint;
 import org.vertexium.util.ConvertingIterable;
 import org.vertexium.util.FilterIterable;
-import org.vertexium.util.IterableUtils;
 
-import java.util.EnumSet;
+import java.util.HashMap;
+import java.util.Map;
 
 public class InMemoryVertex extends InMemoryElement<InMemoryVertex> implements Vertex {
     public InMemoryVertex(
@@ -224,18 +224,35 @@ public class InMemoryVertex extends InMemoryElement<InMemoryVertex> implements V
     }
 
     @Override
+    @Deprecated
     public int getEdgeCount(Direction direction, Authorizations authorizations) {
-        return IterableUtils.count(getEdgeIds(direction, authorizations));
+        return getEdgesSummary(authorizations).getCountOfEdges(direction);
     }
 
     @Override
+    @Deprecated
     public Iterable<String> getEdgeLabels(Direction direction, Authorizations authorizations) {
-        return IterableUtils.toSet(new ConvertingIterable<Edge, String>(getEdges(direction, authorizations)) {
-            @Override
-            protected String convert(Edge o) {
-                return o.getLabel();
-            }
-        });
+        return getEdgesSummary(authorizations).getEdgeLabels(direction);
+    }
+
+    @Override
+    public EdgesSummary getEdgesSummary(Authorizations authorizations) {
+        Map<String, Integer> outEdgeCountsByLabels = new HashMap<>();
+        Map<String, Integer> inEdgeCountsByLabels = new HashMap<>();
+
+        for (EdgeInfo entry : getEdgeInfos(Direction.IN, authorizations)) {
+            String label = entry.getLabel();
+            Integer c = inEdgeCountsByLabels.getOrDefault(label, 0);
+            inEdgeCountsByLabels.put(label, c + 1);
+        }
+
+        for (EdgeInfo entry : getEdgeInfos(Direction.OUT, authorizations)) {
+            String label = entry.getLabel();
+            Integer c = outEdgeCountsByLabels.getOrDefault(label, 0);
+            outEdgeCountsByLabels.put(label, c + 1);
+        }
+
+        return new EdgesSummary(outEdgeCountsByLabels, inEdgeCountsByLabels);
     }
 
     @Override